v0.001

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/rooms/special/ShopRoom.java
@@ -24,6 +24,7 @@ package com.shatteredpixel.shatteredpixeldungeon.levels.rooms.special;
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
 import com.shatteredpixel.shatteredpixeldungeon.ShatteredPixelDungeon;
 import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Belongings;
+import com.shatteredpixel.shatteredpixeldungeon.actors.hero.Hero;
 import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.Mob;
 import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.npcs.Shopkeeper;
 import com.shatteredpixel.shatteredpixeldungeon.items.Ankh;
@@ -63,6 +64,8 @@ import com.watabou.utils.Random;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+
+import network.Multiplayer;
 
 public class ShopRoom extends SpecialRoom {
 
@@ -269,7 +272,7 @@ public class ShopRoom extends SpecialRoom {
 
 		itemsToSpawn.add( new Alchemize().quantity(Random.IntRange(2, 3)));
 
-		Bag bag = ChooseBag(Dungeon.hero.belongings);
+		Bag bag = ChooseBag();
 		if (bag != null) {
 			itemsToSpawn.add(bag);
 		}
@@ -306,28 +309,28 @@ public class ShopRoom extends SpecialRoom {
 
 		itemsToSpawn.add( new Ankh() );
 		itemsToSpawn.add( new StoneOfAugmentation() );
-
-		TimekeepersHourglass hourglass = Dungeon.hero.belongings.getItem(TimekeepersHourglass.class);
-		if (hourglass != null && hourglass.isIdentified() && !hourglass.cursed){
-			int bags = 0;
-			//creates the given float percent of the remaining bags to be dropped.
-			//this way players who get the hourglass late can still max it, usually.
-			switch (Dungeon.depth) {
-				case 6:
-					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.20f ); break;
-				case 11:
-					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.25f ); break;
-				case 16:
-					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.50f ); break;
-				case 20: case 21:
-					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.80f ); break;
-			}
-
-			for(int i = 1; i <= bags; i++){
-				itemsToSpawn.add( new TimekeepersHourglass.sandBag());
-				hourglass.sandBags ++;
-			}
-		}
+		//TODO remake
+//		TimekeepersHourglass hourglass = Dungeon.hero.belongings.getItem(TimekeepersHourglass.class);
+//		if (hourglass != null && hourglass.isIdentified() && !hourglass.cursed){
+//			int bags = 0;
+//			//creates the given float percent of the remaining bags to be dropped.
+//			//this way players who get the hourglass late can still max it, usually.
+//			switch (Dungeon.depth) {
+//				case 6:
+//					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.20f ); break;
+//				case 11:
+//					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.25f ); break;
+//				case 16:
+//					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.50f ); break;
+//				case 20: case 21:
+//					bags = (int)Math.ceil(( 5-hourglass.sandBags) * 0.80f ); break;
+//			}
+//
+//			for(int i = 1; i <= bags; i++){
+//				itemsToSpawn.add( new TimekeepersHourglass.sandBag());
+//				hourglass.sandBags ++;
+//			}
+//		}
 
 		Item rare;
 		switch (Random.Int(10)){
@@ -358,48 +361,44 @@ public class ShopRoom extends SpecialRoom {
 		return itemsToSpawn;
 	}
 
-	protected static Bag ChooseBag(Belongings pack){
-
-		//generate a hashmap of all valid bags.
+	protected static Bag ChooseBag(){
+		// Проверяем доступные сумки (ещё не выпавшие)
 		HashMap<Bag, Integer> bags = new HashMap<>();
-		if (!Dungeon.LimitedDrops.VELVET_POUCH.dropped()) bags.put(new VelvetPouch(), 1);
+		if (!Dungeon.LimitedDrops.VELVET_POUCH.dropped()) bags.put(new VelvetPouch(), 0);
 		if (!Dungeon.LimitedDrops.SCROLL_HOLDER.dropped()) bags.put(new ScrollHolder(), 0);
 		if (!Dungeon.LimitedDrops.POTION_BANDOLIER.dropped()) bags.put(new PotionBandolier(), 0);
 		if (!Dungeon.LimitedDrops.MAGICAL_HOLSTER.dropped()) bags.put(new MagicalHolster(), 0);
 
 		if (bags.isEmpty()) return null;
 
-		//count up items in the main bag
-		for (Item item : pack.backpack.items) {
-			for (Bag bag : bags.keySet()){
-				if (bag.canHold(item)){
-					bags.put(bag, bags.get(bag)+1);
+		// Считаем предметы у всех героев, подходящие для каждой сумки
+		for (Multiplayer.PlayerInfo info : Multiplayer.Players.getAll()) {
+			Hero hero = info.hero;
+			if (hero == null) continue;
+			for (Item item : hero.belongings.backpack.items) {
+				for (Bag bag : bags.keySet()) {
+					if (bag.canHold(item)) {
+						bags.put(bag, bags.get(bag) + 1);
+					}
 				}
 			}
 		}
 
-		//find which bag will result in most inventory savings, drop that.
+		// Выбираем сумку с максимальным количеством подходящих предметов
 		Bag bestBag = null;
-		for (Bag bag : bags.keySet()){
-			if (bestBag == null){
-				bestBag = bag;
-			} else if (bags.get(bag) > bags.get(bestBag)){
+		for (Bag bag : bags.keySet()) {
+			if (bestBag == null || bags.get(bag) > bags.get(bestBag)) {
 				bestBag = bag;
 			}
 		}
 
-		if (bestBag instanceof VelvetPouch){
-			Dungeon.LimitedDrops.VELVET_POUCH.drop();
-		} else if (bestBag instanceof ScrollHolder){
-			Dungeon.LimitedDrops.SCROLL_HOLDER.drop();
-		} else if (bestBag instanceof PotionBandolier){
-			Dungeon.LimitedDrops.POTION_BANDOLIER.drop();
-		} else if (bestBag instanceof MagicalHolster){
-			Dungeon.LimitedDrops.MAGICAL_HOLSTER.drop();
-		}
+		// Отмечаем, что сумка выпала (ограничение на всю игру)
+		if (bestBag instanceof VelvetPouch) Dungeon.LimitedDrops.VELVET_POUCH.drop();
+		else if (bestBag instanceof ScrollHolder) Dungeon.LimitedDrops.SCROLL_HOLDER.drop();
+		else if (bestBag instanceof PotionBandolier) Dungeon.LimitedDrops.POTION_BANDOLIER.drop();
+		else if (bestBag instanceof MagicalHolster) Dungeon.LimitedDrops.MAGICAL_HOLSTER.drop();
 
 		return bestBag;
-
 	}
 
 }
